+++ ThirdPartyAdapters/applovin/CHANGELOG.md
@@ -1,7 +1,29 @@
 # AppLovin Adapter for Google Mobile Ads SDK for Android
 
-## 7.8.5.1
-- Added support for Unity banners.
+## 8.1.3.0
+- Verified compatibility with AppLovin SDK 8.1.3
+
+## 8.1.0.0
+- Verified compatibility with AppLovin SDK 8.1.0
+
+## 8.0.2.1
+- Updated the adapter to invoke the `onRewardedVideoComplete()` ad event.
+
+## 8.0.2.0
+- Verified compatibility with AppLovin SDK 8.0.2
+
+## 8.0.1.1
+- Added support for native ads.
+- Set AdMob as mediation provider on the AppLovin SDK.
+
+## 8.0.1.0
+- Verified compatibility with AppLovin SDK 8.0.1
+
+## 8.0.0.0
+- Verified compatibility with AppLovin SDK 8.0.0
+
+## 7.8.6.0
+- Verified compatibility with AppLovin SDK 7.8.6
 
 ## 7.8.5.0
 - Added support for zones and smart banners.