+++ ThirdPartyAdapters/ironsource/CHANGELOG.md
@@ -1,8 +1,28 @@
 # ironSource Adapter for Google Mobile Ads SDK for Android Changelog
 
+## Version 6.7.11.0
+- Verified compatibility with ironSource SDK version 6.7.11.
+
+## Version 6.7.10.0
+- Verified compatibility with ironSource SDK version 6.7.10.
+
+## Version 6.7.9.1.1
+- Updated the adapter to invoke the `onRewardedVideoComplete()` ad event.
+
+## Version 6.7.9.1.0
+- Verified compatibility with ironSource SDK version 6.7.9.1.
+
+## Version 6.7.9.0
+- Verified compatibility with ironSource SDK version 6.7.9.
+
+## Version 6.7.8.0
+- Verified compatibility with ironSource SDK version 6.7.8.
+
 ## Version 6.7.7.0
-- Initiate IronSource SDK once per adunit
-- Add option to pass activity to IronSource SDK in 'onPause'\'onResume' methods
+- Initialize IronSource SDK once per ad unit.
+- Added option to pass activity to IronSource SDK in `onPause` and `onResume`
+  methods.
+- Verified compatibility with ironSource SDK version 6.7.0.
 
 ## Version 6.7.5.0
-- Initial release!
+- Initial release.