+++ ThirdPartyAdapters/mytarget/CHANGELOG.md
@@ -1,5 +1,46 @@
 # myTarget Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 5.2.2.0
+- Verified compatibility with myTarget SDK version 5.2.2.
+
+## 5.2.1.0
+- Verified compatibility with myTarget SDK version 5.2.1.
+
+## 5.2.0.0
+- Verified compatibility with myTarget SDK version 5.2.0.
+
+## 5.1.4.0
+- Verified compatibility with myTarget SDK version 5.1.4.
+
+## 5.1.3.0
+- Verified compatibility with myTarget SDK version 5.1.3.
+
+## 5.1.2.0
+- Verified compatibility with myTarget SDK version 5.1.2.
+
+## 5.1.1.0
+- Verified compatibility with myTarget SDK version 5.1.1.
+
+## 5.1.0.1
+- Updated adapter to use the Unified Native Ads adapter API.
+
+## 5.1.0.0
+- Verified compatibility with myTarget SDK version 5.1.0.
+
+## 5.0.4.0
+- Verified compatibility with myTarget SDK version 5.0.4.
+
+## 5.0.2.1
+- Updated the adapter to make it compatible with Google Mobile Ads SDK version 15.0.0.
+
+## 5.0.2.0
+- Verified compatibility with myTarget SDK version 5.0.2.
+
+## 5.0.0.0
+- Verified compatibility with myTarget SDK version 5.0.0.
+- Updated the adapter to make it compatible with myTarget SDK 5.0.0 and Android
+  Studio 3.1.
+
 ## 4.7.2.0
 - Verified compatibility with myTarget SDK version 4.7.2.