+++ ThirdPartyAdapters/nend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Nend Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 5.0.2.0
+
+- Verified compatibility with nend SDK 5.0.2.
+
+## 5.0.1.0
+
+- Verified compatibility with nend SDK 5.0.1.
+
+## 5.0.0.0
+
+- Verified compatibility with nend SDK 5.0.0.
+
+## 4.0.5.0
+
+- Verified compatibility with nend SDK 4.0.5.
+
 ## 4.0.4.1
 
 - Supported onRewardedVideoCompleted() method of RewardedVideoAdListener.