+++ ThirdPartyAdapters/tapjoy/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Tapjoy Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 12.0.0.1
+- Updated the adapter to invoke the `onRewardedVideoComplete()` ad event.
+
+## 12.0.0.0
+- Verified compatibility with tapjoy SDK versiobn 12.0.0.
+
+## 11.12.2.0
+- Verified compatibility with tapjoy SDK versiobn 11.12.2.
+
+## 11.12.1.0
+- Verified compatibility with tapjoy SDK versiobn 11.12.1.
+
+## 11.12.0.0
+- Verified compatibility with tapjoy SDK versiobn 11.12.0.
+
 ## 11.11.1.0
 - Verified compatibility with Tapjoy SDK version 11.11.1.